+++ itest/types/queries.go
@@ -1,0 +1,46 @@
+package types
+
+type ConsumerFpsResponse struct {
+	ConsumerFps []SingleConsumerFpResponse `json:"fps"`
+}
+
+// SingleConsumerFpResponse represents the finality provider data returned by the contract query.
+// For more details, refer to the following links:
+// https://github.com/babylonchain/babylon-contract/blob/v0.5.3/packages/apis/src/btc_staking_api.rs
+// https://github.com/babylonchain/babylon-contract/blob/v0.5.3/contracts/btc-staking/src/msg.rs
+// https://github.com/babylonchain/babylon-contract/blob/v0.5.3/contracts/btc-staking/schema/btc-staking.json
+type SingleConsumerFpResponse struct {
+	BtcPkHex             string `json:"btc_pk_hex"`
+	SlashedBabylonHeight uint64 `json:"slashed_babylon_height"`
+	SlashedBtcHeight     uint64 `json:"slashed_btc_height"`
+	ConsumerId           string `json:"consumer_id"`
+}
+
+type ConsumerDelegationsResponse struct {
+	ConsumerDelegations []SingleConsumerDelegationResponse `json:"delegations"`
+}
+
+type SingleConsumerDelegationResponse struct {
+	BtcPkHex             string                      `json:"btc_pk_hex"`
+	FpBtcPkList          []string                    `json:"fp_btc_pk_list"`
+	StartHeight          uint64                      `json:"start_height"`
+	EndHeight            uint64                      `json:"end_height"`
+	TotalSat             uint64                      `json:"total_sat"`
+	StakingTx            []byte                      `json:"staking_tx"`
+	SlashingTx           []byte                      `json:"slashing_tx"`
+	DelegatorSlashingSig []byte                      `json:"delegator_slashing_sig"`
+	CovenantSigs         []CovenantAdaptorSignatures `json:"covenant_sigs"`
+	StakingOutputIdx     uint32                      `json:"staking_output_idx"`
+	UnbondingTime        uint32                      `json:"unbonding_time"`
+	UndelegationInfo     *BtcUndelegationInfo        `json:"undelegation_info"`
+	ParamsVersion        uint32                      `json:"params_version"`
+}
+
+type ConsumerFpInfo struct {
+	BtcPkHex string `json:"btc_pk_hex"`
+	Power    uint64 `json:"power"`
+}
+
+type ConsumerFpsByPowerResponse struct {
+	Fps []ConsumerFpInfo `json:"fps"`
+}
